+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Updated `PdfFont` constructor.
 - Updated `StyleCI` configuration.
 - Updated `PHP-CS-Fixer` to version 3.88.0 and the configuration.